turn carveout into library
