+++ Cargo.toml
@@ -44,3 +44,6 @@ opt-level = 1
 # Enable high optimizations for dependencies
 [profile.dev.package."*"]
 opt-level = 3
+
+[lib]
+crate-type = [ "rlib", "cdylib" ]
